+++ modules/core/src/test/scala/org/scalasteward/core/edit/scalafix/ScalafixMigrationsLoaderTest.scala
@@ -5,11 +5,9 @@ import munit.FunSuite
 import org.http4s.Uri
 import org.scalasteward.core.application.Config.ScalafixCfg
 import org.scalasteward.core.data.{GroupId, Version}
-import org.scalasteward.core.edit.scalafix.ScalafixMigrationsLoaderTest.mockState
-import org.scalasteward.core.io.FileAlgTest.ioFileAlg
 import org.scalasteward.core.mock.MockConfig.mockRoot
 import org.scalasteward.core.mock.MockContext.context.scalafixMigrationsLoader
-import org.scalasteward.core.mock.MockState
+import org.scalasteward.core.mock.MockContext.mockState
 import org.scalasteward.core.util.Nel
 
 class ScalafixMigrationsLoaderTest extends FunSuite {
@@ -77,11 +75,4 @@ class ScalafixMigrationsLoaderTest extends FunSuite {
       .unsafeRunSync()
     assert(migrations.isLeft)
   }
-}
-
-object ScalafixMigrationsLoaderTest {
-  val mockState: MockState = MockState.empty.addUris(
-    ScalafixMigrationsLoader.defaultScalafixMigrationsUrl ->
-      ioFileAlg.readResource("scalafix-migrations.conf").unsafeRunSync()
-  )
 }
